Refactor Snoop to Use Config
Move to using the Config object in the Snoop class. This required minimal
changes to the unit tests of the snoop module, mainly to update the accesses to
the members of the config class rather than the snooper object itself. This is
not a breaking API change as we are hacking the access to these members in
anyhow just for the unit tests.

During the process it was found that the Config object didn't translate things
into Hooks. This is now implemented.

Behold the lightness of the Snoop classs with regards to config option handling.